+++ app/src/main/java/blog/rishabh/coroutinesretrofitdemo/model/ApiAdapter.kt
@@ -1,4 +1,4 @@
-package blog.rishabh.coroutinesretrofitdemo
+package blog.rishabh.coroutinesretrofitdemo.model
 
 import okhttp3.OkHttpClient
 import retrofit2.Retrofit
