+++ app/src/main/java/blog/rishabh/coroutinesretrofitdemo/model/DogImageModel.kt
@@ -1,4 +1,4 @@
-package blog.rishabh.coroutinesretrofitdemo
+package blog.rishabh.coroutinesretrofitdemo.model
 
 import androidx.annotation.Keep
